refactor(test): merge test_reddit into test_scrapers_reddit
合并 2 个唯一测试（浏览器头信息和 403 评论降级）至 test_scrapers_reddit.py，删除旧文件
